paid_server/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
argcomplete==3.1.2
async-timeout==4.0.3
attrs==24.2.0
base58==2.1.1
bitarray==3.0.0
certifi==2024.8.30
charset-normalizer==3.4.0
ckzg==2.0.1
click==8.1.7
cytoolz==1.0.0
dnspython==2.7.0
email_validator==2.2.0
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==0.3.0
eth-typing==5.0.1
eth-utils==2.3.1
eth_abi==5.1.0
exceptiongroup==1.2.2
fastapi==0.115.2
fastapi-cli==0.0.5
frozenlist==1.4.1
grpcio==1.67.0
grpcio-health-checking==1.59.0
grpcio-tools==1.59.0
h11==0.14.0
hexbytes==0.3.1
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
ipfshttpclient==0.4.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
jsonrpcclient==4.0.3
jsonschema==4.0.0
lighthouseweb3==0.1.4
lru-dict==1.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdurl==0.1.2
mnemonic==0.20
multiaddr==0.0.9
multidict==6.1.0
netaddr==1.3.0
orjson==3.10.7
parsimonious==0.10.0
propcache==0.2.0
protobuf==4.21.6
pycoin==0.92.20230326
pycryptodome==3.21.0
pydantic==2.9.2
pydantic-extra-types==2.9.0
pydantic-settings==2.6.0
pydantic_core==2.23.4
Pygments==2.18.0
pymultihash==0.8.2
pyrsistent==0.20.0
python-dotenv==1.0.1
python-multipart==0.0.12
pyunormalize==16.0.0
PyYAML==6.0.1
regex==2024.9.11
requests==2.32.3
rfc3986==2.0.0
rich==13.9.2
rlp==3.0.0
shellingham==1.5.4
six==1.16.0
snet.contracts==0.1.1
snet.sdk==3.6.0
sniffio==1.3.1
starlette==0.40.0
toolz==1.0.0
typer==0.12.5
types-requests==2.32.0.20241016
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
varint==1.0.2
watchfiles==0.24.0
web3==6.11.1
websockets==13.1
yarl==1.15.4
zipp==3.20.2
